Страница добавления вещи перенесена в отдельный роут
+ мелкие исправления
