+++ frontend/src/js/hooks/use-show-items-request-error.ts
@@ -1,10 +1,10 @@
 import { useEffect } from 'react';
 import UIkit from 'uikit';
 import { useSelector } from 'react-redux';
-import { itemsErrorSelector, userRequestErrorSelector } from '#redux/selectors';
+import { itemsRequestErrorSelector, userRequestErrorSelector } from '#redux/selectors';
 
 export const useShowItemsRequestError = (isSubmitSuccessful: boolean): void => {
-  const itemsRequestError = useSelector(itemsErrorSelector);
+  const itemsRequestError = useSelector(itemsRequestErrorSelector);
 
   useEffect(() => {
     if (isSubmitSuccessful && itemsRequestError)
